tracking_wo_bnw/requirements.txt
absl-py==0.11.0
astroid==2.4.2
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
cached-property==1.5.2
cachetools==3.1.1
certifi==2020.12.5
chardet==4.0.0
clang==5.0
colorama==0.4.1
cycler==0.10.0
Cython==0.29.16
dataclasses==0.6
decorator==4.4.1
defusedxml==0.6.0
docopt==0.6.2
easydict==1.9
entrypoints==0.3
enum34==1.1.6
fastdtw==0.3.4
filelock==3.0.12
flake8==3.9.1
flake8-import-order==0.18.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gdown==3.12.2
gitdb2==2.0.6
GitPython==3.0.5
google==2.0.2
google-auth==1.26.1
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.39.0
h5py==3.1.0
idna==2.10
imageio==2.9.0
importlib-metadata==0.23
iniconfig==1.1.1
ipykernel==5.1.3
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.3
jsonpickle==0.9.6
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyter-http-over-ws==0.0.6
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
lap==0.4.0
lapsolver==1.0.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
matlab==0.1
matlab-kernel==0.16.11
# matlabengineforpython===R2021a
matplotlib==3.1.1
mccabe==0.6.1
metakernel==0.27.5
mistune==0.8.4
more-itertools==7.2.0
motmetrics==1.2.0
munch==2.5.0
munkres==1.1.2
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.5.1
notebook==6.0.2
numpy>=1.11
oauthlib==3.1.0
opencv-python==4.0.1.24
opt-einsum==3.3.0
packaging==19.2
pandas==1.3.4
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.4.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.2
ptyprocess==0.6.0
py==1.10.0
py-cpuinfo==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycocotools==2.0.7
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.4.2
pylint==2.6.0
pyparsing==2.4.5
pyrsistent==0.15.5
PySocks==1.7.1
pytest==6.2.4
pytest-benchmark==3.4.1
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==18.1.1
qtconsole==4.5.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.0
sacred==0.8.0
scikit-image==0.18.1
scipy==1.3.2
seaborn==0.11.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.15.0
smmap2==2.0.5
soupsieve==1.9.5
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tifffile==2021.4.8
toml==0.10.2
torchreid @ https://github.com/KaiyangZhou/deep-person-reid/archive/master.zip
tornado==6.0.4
tqdm==4.38.0
traitlets==4.3.3
typed-ast==1.4.1
#typing-extensions==3.10
urllib3==1.26.3
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wurlitzer==3.0.2
xmltodict==0.12.0
yacs==0.1.8
yapf==0.31.0
zipp==0.6.0
#for ssl
psutil==5.9.8
imutils==0.5.4
scikit-learn==1.0.2
